VisionInferenceServer/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
Pillow
loguru
requests
ImageHash
